+++ src/main/java/com/cotato/weather/domain/place/dto/response/DailyResponse.java
@@ -3,10 +3,11 @@ package com.cotato.weather.domain.place.dto.response;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
+import lombok.NoArgsConstructor;
 
 @Getter
 @Builder
-@AllArgsConstructor
+@AllArgsConstructor @NoArgsConstructor
 public class DailyResponse {
     private String date;
     private String weatherDescription;
